refactor: extract shared API wire-contract types to @wow-threat/shared
Create packages/shared/src/api-types.ts with all shared HTTP response types,
replacing near-identical duplicates in apps/api/src/types/api.ts and
apps/web/src/types/api.ts. Both files now re-export from @wow-threat/shared.

Also fixes zone: Zone | null mismatch in web copy (now Zone, matching API contract).
AugmentedEventsResponse remains web-only in apps/web/src/types/api.ts.
