+++ packages/shared/src/index.ts
@@ -2,6 +2,7 @@ export * from './utils'
 export * from './types'
 export * from './cache-versions'
 export * from './aura-serialization'
+export * from './api-types'
 export * from './test/helpers/context'
 export * from './test/helpers/config'
 export * from './test/helpers/events'
